Rename engine files to avoid conflicts with common files
