+++ SymCryptEngine/src/e_scossl_rsa.h
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft Corporation. Licensed under the MIT license.
 //
 
-#include "scossl.h"
+#include "e_scossl.h"
 #include <openssl/rsa.h>
 
 #ifdef __cplusplus
