+++ SymCryptEngine/src/e_scossl_rsapss.h
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft Corporation. Licensed under the MIT license.
 //
 
-#include "scossl.h"
+#include "e_scossl.h"
 
 #ifdef __cplusplus
 extern "C" {
